chore: split node-api bindings into separate file
This helps ensure that the core logic does not accidentally interact
with JS objects.
