+++ package.json
@@ -32,7 +32,7 @@
     "win32"
   ],
   "files": [
-    "src/binding.cc",
+    "src",
     "binding.gyp",
     "index.d.ts",
     "index.js",
